Move downloadTrack to lib module.
