+++ lib/index.js
@@ -1,5 +1,6 @@
 const getBody = require('./getBody');
 const getTracks = require('./getTracks');
 const validateTracks = require('./validateTracks');
+const downloadTrack = require('./downloadTrack');
 
-module.exports = { getBody, getTracks, validateTracks };
+module.exports = { getBody, getTracks, validateTracks, downloadTrack };
